refactor: split useQuery code into separate files
