+++ src/query/useRecommendations.ts
@@ -1,0 +1,11 @@
+import { getRecommendations } from '@/api/spotify';
+import { useQuery } from '@tanstack/react-query';
+
+export const useRecommendations = () => {
+  return useQuery({
+    queryKey: useRecommendations.queryKey(),
+    queryFn: () => getRecommendations(),
+  });
+};
+
+useRecommendations.queryKey = () => ['recommendations'];
